Moved authorization testing endpoints from datasets router to authorizations router.
